rename all fiels
