+++ crawler/package.json
@@ -25,7 +25,7 @@
     "divinator": "^1.0.1",
     "dotenv": "16.4.7",
     "is-valid-domain": "^0.1.6",
-    "node-cron": "^3.0.2",
+    "node-cron": "^3.0.3",
     "pm2": "5.4.3",
     "redis": "4.7.0",
     "remove-markdown": "0.6.0",
@@ -34,6 +34,7 @@
   },
   "devDependencies": {
     "@types/node": "22.10.3",
+    "@types/node-cron": "^3.0.11",
     "tslib": "2.8.1",
     "typescript": "5.7.2"
   }
