+++ crawler/src/queue/check_comm.ts
@@ -1,12 +1,12 @@
 import logging from "../lib/logging";
 
-import storage from "../storage.js";
+import storage from "../storage";
 
-import { CrawlTooRecentError } from "../lib/error.js";
+import { CrawlTooRecentError } from "../lib/error";
 
-import CommunityCrawler from "../crawl/community.js";
+import CommunityCrawler from "../crawl/community";
 
-import BaseQueue from "./queue.js";
+import BaseQueue from "./queue";
 
 export default class SingleCommunityQueue extends BaseQueue {
   constructor(isWorker = false, queueName = "one_community") {
