+++ crawler/src/storage/instance.ts
@@ -1,3 +1,5 @@
+import Storage from "../storage";
+
 /**
  * Stores each lemmy instance, keyed on baseUrl as `instance:baseUrl`.
  *
@@ -12,21 +14,28 @@
 // };
 
 export default class Instance {
-  constructor(storage) {
+  private storage: Storage;
+
+  constructor(storage: Storage) {
     this.storage = storage;
   }
+
   async getAll() {
     return this.storage.listRedis(`instance:*`);
   }
+
   async getAllWithKeys() {
     return this.storage.listRedisWithKeys(`instance:*`);
   }
+
   async getOne(key) {
     return this.storage.getRedis(`instance:${key}`);
   }
+
   async upsert(baseUrl, value) {
     return this.storage.putRedis(`instance:${baseUrl}`, value);
   }
+
   async delete(key) {
     return this.storage.deleteRedis(`instance:${key}`);
   }
